事务执行

+++ common.go
@@ -86,8 +86,8 @@ func PathExists(path string) (bool, error) {
 }
 
 // 将时间缀格式为YYYY-MM-DD HH:mm:ss
-func FormatTime(i int) string {
-	return time.Unix(int64(i), 0).Format("2006-01-02 :15:04:05")
+func FormatTime(i int64) string {
+	return time.Unix(i, 0).Format("2006-01-02 :15:04:05")
 }
 
 // 返回工作目录
